+++ src/main/java/baseball/controller/NumberComparator.java
@@ -1,7 +1,7 @@
 package baseball.controller;
 
 public class NumberComparator {
-    public static boolean isEqual(int computerNumber, int playerNumber) {
+    static boolean isEqual(int computerNumber, int playerNumber) {
         return computerNumber == playerNumber;
     }
 }
