+++ src/test/java/baseball/controller/NumberComparatorTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.controller;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
